LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/dynamic-import/dynamic-imports-credentials-setTimeout.sub-expected.txt



PASS Same-origin dynamic import from setTimeout(string) from classic script (crossOrigin not specified)
FAIL Cross-origin dynamic import from setTimeout(string) from classic script (crossOrigin not specified) assert_equals: cookie expected "not found" but got "found"
PASS Same-origin dynamic import from setTimeout(string) from classic script (crossOrigin=anonymous)
FAIL Cross-origin dynamic import from setTimeout(string) from classic script (crossOrigin=anonymous) assert_equals: cookie expected "not found" but got "found"
PASS Same-origin dynamic import from setTimeout(string) from classic script (crossOrigin=use-credentials)
PASS Cross-origin dynamic import from setTimeout(string) from classic script (crossOrigin=use-credentials)
PASS Same-origin dynamic import from setTimeout(string) from module script (crossOrigin not specified)
FAIL Cross-origin dynamic import from setTimeout(string) from module script (crossOrigin not specified) assert_equals: cookie expected "not found" but got "found"
PASS Same-origin dynamic import from setTimeout(string) from module script (crossOrigin=anonymous)
FAIL Cross-origin dynamic import from setTimeout(string) from module script (crossOrigin=anonymous) assert_equals: cookie expected "not found" but got "found"
PASS Same-origin dynamic import from setTimeout(string) from module script (crossOrigin=use-credentials)
PASS Cross-origin dynamic import from setTimeout(string) from module script (crossOrigin=use-credentials)
